crops_research_gosha20777/requirements.txt
jupyter
opencv-python
tensorflow-gpu
keras==2.3.1
click
tqdm
joblib
matplotlib
nympy
sklearn
Pillow